Move resourceutil from pkg/ into internal/pkg
