+++ fleetshard/pkg/central/reconciler/reconciler.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stackrox/acs-fleet-manager/fleetshard/pkg/util"
 	centralConstants "github.com/stackrox/acs-fleet-manager/internal/dinosaur/constants"
 	"github.com/stackrox/acs-fleet-manager/internal/dinosaur/pkg/api/private"
-	api "github.com/stackrox/acs-fleet-manager/pkg/api"
+	"github.com/stackrox/acs-fleet-manager/internal/dinosaur/pkg/converters"
 	"github.com/stackrox/rox/operator/apis/platform/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -70,19 +70,19 @@ func (r *CentralReconciler) Reconcile(ctx context.Context, remoteCentral private
 
 	centralMonitoringExposeEndpointEnabled := v1alpha1.ExposeEndpointEnabled
 
-	centralResources, err := api.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Central.Resources)
+	centralResources, err := converters.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Central.Resources)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting Central resources")
 	}
-	scannerAnalyzerResources, err := api.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Scanner.Analyzer.Resources)
+	scannerAnalyzerResources, err := converters.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Scanner.Analyzer.Resources)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting Scanner Analyzer resources")
 	}
-	scannerAnalyzerScaling, err := api.ConvertPrivateScalingToV1(&remoteCentral.Spec.Scanner.Analyzer.Scaling)
+	scannerAnalyzerScaling, err := converters.ConvertPrivateScalingToV1(&remoteCentral.Spec.Scanner.Analyzer.Scaling)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting Scanner Scaling resources")
 	}
-	scannerDbResources, err := api.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Scanner.Db.Resources)
+	scannerDbResources, err := converters.ConvertPrivateResourceRequirementsToCoreV1(&remoteCentral.Spec.Scanner.Db.Resources)
 	if err != nil {
 		return nil, errors.Wrap(err, "converting Scanner DB resources")
 	}
